refactor(dccpeep): extract loop registerization passes
Move the seven loop-scoped BC/C/E/IYL registerization passes and their private zero-extension/IY/escape proof helpers into peep_pass_loops.c. Keep shared label-range, loop-back, internal-label, and jump-target analysis in the private optimizer API because later stride/walk passes use the same proofs.

dccpeep.c drops from 9,200 to about 7,900 lines. Validation: strict build clean; all 316 raw inputs byte-identical in default, -Os, and -fundocumented-z80 modes; six direct fixtures pass.
